Adapt to new project structure
