--- prometheus-source-flink/src/main/java/com/github/ivanas93/example/reader/model/Label.java
@@ -1,13 +1,0 @@
-package com.github.ivanas93.example.reader.model;
-
-import lombok.Builder;
-import lombok.ToString;
-
-import java.io.Serializable;
-
-@Builder
-@ToString
-public class Label implements Serializable {
-    private String name;
-    private String value;
-}
