--- prometheus-source-flink/src/main/java/com/github/ivanas93/example/reader/model/TimeSerie.java
@@ -1,14 +1,0 @@
-package com.github.ivanas93.example.reader.model;
-
-import lombok.Builder;
-import lombok.ToString;
-
-import java.io.Serializable;
-import java.util.List;
-
-@Builder
-@ToString
-public class TimeSerie implements Serializable {
-    private List<Label> labels;
-    private List<Sample> samples;
-}
