+++ prometheus-source-flink/src/main/java/com/github/ivanas93/reader/configuration/PrometheusConfiguration.java
@@ -1,4 +1,4 @@
-package com.github.ivanas93.example.reader.configuration;
+package com.github.ivanas93.reader.configuration;
 
 import lombok.Getter;
