+++ prometheus-source-flink/src/test/java/com/github/ivanas93/reader/test/SnappyContentUtil.java
